Update readme.md
There's too much in this lecture for the morning only and Object.create has a lot of nuances that we don't have time to delve into at this point.  Removing it so we don't get caught up in the conversation of "when do I use this?".  We cover prototypes next and can touch upon it during that lecture, if there's time.